refactor: Push export logic into `ExportMenu` component
