+++ web-common/src/features/dashboards/dimension-table/DimensionHeader.svelte
@@ -15,10 +15,6 @@
   import DelayedSpinner from "@rilldata/web-common/features/entity-management/DelayedSpinner.svelte";
   import { featureFlags } from "@rilldata/web-common/features/feature-flags";
   import { slideRight } from "@rilldata/web-common/lib/transitions";
-  import {
-    V1ExportFormat,
-    createQueryServiceExport,
-  } from "@rilldata/web-common/runtime-client";
   import { onDestroy } from "svelte";
   import { fly } from "svelte/transition";
   import ExportMenu from "../../exports/ExportMenu.svelte";
@@ -26,7 +22,6 @@
   import { getStateManagers } from "../state-managers/state-managers";
   import SelectAllButton from "./SelectAllButton.svelte";
   import { getDimensionTableExportArgs } from "./dimension-table-export-utils";
-  import exportToplist from "./export-toplist";
 
   export let dimensionName: string;
   export let isFetching: boolean;
@@ -35,8 +30,6 @@
   export let searchText: string;
   export let onToggleSearchItems: () => void;
   export let hideStartPivotButton = false;
-
-  const exportDash = createQueryServiceExport();
 
   const stateManagers = getStateManagers();
   const {
@@ -139,15 +132,6 @@
 
   const scheduledReportsQueryArgs = getDimensionTableExportArgs(stateManagers);
 
-  const handleExportTopList = async (format: V1ExportFormat) => {
-    await exportToplist({
-      ctx: stateManagers,
-      query: exportDash,
-      format,
-      searchText,
-    });
-  };
-
   onDestroy(() => {
     searchText = "";
   });
@@ -216,11 +200,12 @@
     {#if $exports}
       <ExportMenu
         label="Export dimension table data"
-        onExport={handleExportTopList}
         includeScheduledReport={$adminServer}
-        queryArgs={$scheduledReportsQueryArgs}
-        {metricsViewProto}
         exploreName={$exploreName}
+        query={{
+          metricsViewAggregationRequest: $scheduledReportsQueryArgs,
+        }}
+        {metricsViewProto}
       />
     {/if}
     {#if !hideStartPivotButton}
