+++ web-common/src/features/dashboards/pivot/PivotToolbar.svelte
@@ -4,14 +4,10 @@
   import { metricsExplorerStore } from "@rilldata/web-common/features/dashboards/stores/dashboard-stores";
   import Spinner from "@rilldata/web-common/features/entity-management/Spinner.svelte";
   import { EntityStatus } from "@rilldata/web-common/features/entity-management/types";
+  import ExportMenu from "../../exports/ExportMenu.svelte";
   import { featureFlags } from "../../feature-flags";
   import { getStateManagers } from "../state-managers/state-managers";
-  import ExportMenu from "../../exports/ExportMenu.svelte";
-  import {
-    V1ExportFormat,
-    createQueryServiceExport,
-  } from "../../../runtime-client";
-  import exportPivot, { getPivotExportArgs } from "./pivot-export";
+  import { getPivotExportArgs } from "./pivot-export";
 
   export let showPanels = true;
   export let isFetching = false;
@@ -19,7 +15,7 @@
   const { adminServer, exports } = featureFlags;
 
   const stateManagers = getStateManagers();
-  const { exploreName, dashboardStore, validSpecStore } = stateManagers;
+  const { exploreName, dashboardStore } = stateManagers;
 
   $: expanded = $dashboardStore?.pivot?.expanded ?? {};
   $: metricsViewProto = $dashboardStore.proto;
@@ -54,17 +50,6 @@
   // }
 
   const scheduledReportsQueryArgs = getPivotExportArgs(stateManagers);
-
-  const exportDash = createQueryServiceExport();
-
-  async function handleExportPivot(format: V1ExportFormat) {
-    await exportPivot({
-      ctx: stateManagers,
-      query: exportDash,
-      format,
-      timeDimension: $validSpecStore.data?.metricsView?.timeDimension,
-    });
-  }
 </script>
 
 <div class="flex items-center gap-x-4 select-none pointer-events-none">
@@ -108,10 +93,11 @@
   {#if $exports}
     <ExportMenu
       label="Export pivot data"
-      onExport={handleExportPivot}
       includeScheduledReport={$adminServer}
-      queryArgs={$scheduledReportsQueryArgs}
       exploreName={$exploreName}
+      query={{
+        metricsViewAggregationRequest: $scheduledReportsQueryArgs,
+      }}
       {metricsViewProto}
     />
   {/if}
