+++ web-common/src/features/models/workspace/ModelWorkspaceCTAs.svelte
@@ -6,11 +6,9 @@
   import CaretDownIcon from "@rilldata/web-common/components/icons/CaretDownIcon.svelte";
   import MetricsViewIcon from "@rilldata/web-common/components/icons/MetricsViewIcon.svelte";
   import { removeLeadingSlash } from "@rilldata/web-common/features/entity-management/entity-mappers";
-  import { createExportTableMutation } from "@rilldata/web-common/features/models/workspace/export-table";
   import { BehaviourEventMedium } from "@rilldata/web-common/metrics/service/BehaviourEventTypes";
   import { MetricsEventSpace } from "@rilldata/web-common/metrics/service/MetricsTypes";
   import {
-    V1ExportFormat,
     V1ReconcileStatus,
     type V1Resource,
   } from "@rilldata/web-common/runtime-client";
@@ -30,8 +28,6 @@
   export let hasUnsavedChanges: boolean;
   export let connector: string;
 
-  const exportModelMutation = createExportTableMutation();
-
   $: ({ instanceId } = $runtime);
   $: isModelIdle =
     resource?.meta?.reconcileStatus === V1ReconcileStatus.RECONCILE_STATUS_IDLE;
@@ -50,16 +46,6 @@
     BehaviourEventMedium.Menu,
     MetricsEventSpace.LeftPanel,
   );
-
-  const onExport = async (format: V1ExportFormat) => {
-    return $exportModelMutation.mutateAsync({
-      data: {
-        instanceId,
-        format,
-        tableName: modelName,
-      },
-    });
-  };
 </script>
 
 <ModelRefreshButton {resource} {hasUnsavedChanges} />
@@ -67,7 +53,12 @@
 <ExportMenu
   label="Export model data"
   disabled={modelHasError || !isModelIdle}
-  {onExport}
+  query={{
+    tableRowsRequest: {
+      instanceId,
+      tableName: modelName,
+    },
+  }}
   workspace
 />
